refactor: 💡 tradingChartSwitch, tradingView
