+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "src",
-  "version": "0.8.0+33.4",
+  "version": "0.8.0+33.5",
   "private": true,
   "scripts": {
     "dev": "next dev",
