--- src/components/position_tab_mobile/position_tab_mobile.tsx
@@ -1,42 +1,0 @@
-import React, {Dispatch, SetStateAction} from 'react';
-import Image from 'next/image';
-import {ImCross} from 'react-icons/im';
-import OpenSubTab from '../open_sub_tab/open_sub_tab';
-import HistorySubTab from '../history_sub_tab/history_sub_tab';
-
-interface PositionTabMobileProps {
-  showSubMenu: boolean;
-  setShowSubMenu: Dispatch<SetStateAction<boolean>>;
-  activeTab: string;
-}
-
-const PositionTabMobile = ({showSubMenu, setShowSubMenu, activeTab}: PositionTabMobileProps) => {
-  const subMenuHandler = () => setShowSubMenu(!showSubMenu);
-
-  const currentSubTab = activeTab === 'Open' ? <OpenSubTab /> : <HistorySubTab />;
-
-  const subMenu = (
-    <div
-      className={`flex h-screen w-screen flex-col items-center justify-center bg-darkGray ${
-        showSubMenu ? 'visible translate-y-0 opacity-100' : 'invisible translate-y-full opacity-0'
-      } absolute left-0 ${'bottom-76px'} overflow-hidden pt-40 transition-all duration-150`}
-    >
-      <div className="mb-3 mr-30px flex self-end sm:pr-30px">
-        <ImCross onClick={subMenuHandler} className="cursor-pointer" />
-      </div>
-      {currentSubTab}
-    </div>
-  );
-
-  return (
-    <div className="rounded-md bg-darkGray5 p-3">
-      <button onClick={subMenuHandler} className="m-auto block">
-        <Image src="/elements/position_tab_icon.svg" alt="position_icon" width={25} height={25} />
-      </button>
-
-      {subMenu}
-    </div>
-  );
-};
-
-export default PositionTabMobile;
